added some changes in Dashboard and ProjectSection
